refactor: replace `additional-components` package occurrences
