+++ docs/components/DocsRepl.vue
@@ -12,6 +12,8 @@ const { vueFlowVersion } = useVueFlow()
 
 let css = `@import 'https://cdn.jsdelivr.net/npm/@vue-flow/core@${vueFlowVersion.value}/dist/style.css';
 @import 'https://cdn.jsdelivr.net/npm/@vue-flow/core@${vueFlowVersion.value}/dist/theme-default.css';
+@import 'https://cdn.jsdelivr.net/npm/@vue-flow/controls@latest/dist/style.css';
+@import 'https://cdn.jsdelivr.net/npm/@vue-flow/minimap@latest/dist/style.css';
 
 html,
 body,
@@ -72,7 +74,9 @@ onMounted(async () => {
   // pre-set import map
   await store.setImportMap({
     imports: {
-      '@vue-flow/additional-components': `${location.origin}/vue-flow-additional-components.mjs`,
+      '@vue-flow/background': `${location.origin}/vue-flow-background.mjs`,
+      '@vue-flow/controls': `${location.origin}/vue-flow-controls.mjs`,
+      '@vue-flow/minimap': `${location.origin}/vue-flow-minimap.mjs`,
       '@vue-flow/core': `${location.origin}/vue-flow-core.mjs`,
       '@vue-flow/node-toolbar': `${location.origin}/vue-flow-node-toolbar.mjs`,
       ...additionalImports,
