+++ docs/package.json
@@ -16,7 +16,10 @@
     "@animxyz/vue3": "^0.6.7",
     "@stackblitz/sdk": "^1.8.0",
     "@vercel/analytics": "^0.1.5",
+    "@vue-flow/background": "workspace:*",
+    "@vue-flow/controls": "workspace:*",
     "@vue-flow/core": "workspace:*",
+    "@vue-flow/minimap": "workspace:*",
     "@vue-flow/node-toolbar": "workspace:*",
     "@vue/repl": "1.1.2",
     "blobity": "^0.2.1",
