+++ docs/src/.vitepress/theme/index.ts
@@ -7,6 +7,8 @@ import 'virtual:windi.css'
 import '@animxyz/core'
 import '@vue-flow/core/dist/style.css'
 import '@vue-flow/core/dist/theme-default.css'
+import '@vue-flow/controls/dist/style.css'
+import '@vue-flow/minimap/dist/style.css'
 import './../../assets/index.css'
 
 const CustomTheme = {
