+++ docs/typedoc.md.json
@@ -7,7 +7,9 @@
   ],
   "entryPoints": [
     "../packages/core/src/index.ts",
-    "../packages/additional-components/src/index.ts",
+    "../packages/background/src/index.ts",
+    "../packages/controls/src/index.ts",
+    "../packages/minimap/src/index.ts",
     "../packages/node-toolbar/src/index.ts"
   ],
   "allReflectionsHaveOwnDocument": true,
