+++ e2e/cypress/component/3-additional-components/background/App.vue
@@ -1,6 +1,6 @@
 <script setup>
 import { VueFlow } from '@vue-flow/core'
-import { Background } from '@vue-flow/additional-components'
+import { Background } from '@vue-flow/background'
 </script>
 
 <template>
