+++ examples/vite/src/EdgeTypes/EdgeTypesExample.vue
@@ -1,19 +1,15 @@
 <script lang="ts" setup>
-import type { VueFlowStore } from '@vue-flow/core'
 import { VueFlow } from '@vue-flow/core'
-import { Controls, MiniMap } from '@vue-flow/additional-components'
+import { Controls } from '@vue-flow/controls'
+import { MiniMap } from '@vue-flow/minimap'
+
 import { getElements } from './utils'
 
-const onLoad = (flowInstance: VueFlowStore) => {
-  flowInstance.fitView()
-  console.log(flowInstance.getElements.value)
-}
-
-const elements = getElements()
+const elements = ref(getElements())
 </script>
 
 <template>
-  <VueFlow v-model="elements" :min-zoom="0.2" @pane-ready="onLoad">
+  <VueFlow v-model="elements" :min-zoom="0.2" fit-view-on-init>
     <MiniMap />
     <Controls />
   </VueFlow>
