+++ packages/vue-flow/src/index.ts
@@ -1,4 +1,9 @@
 import '@vue-flow/core/dist/style.css'
 import '@vue-flow/core/dist/theme-default.css'
-export * from '@vue-flow/additional-components'
+import '@vue-flow/controls/dist/style.css'
+import '@vue-flow/minimap/dist/style.css'
+
+export * from '@vue-flow/background'
+export * from '@vue-flow/controls'
+export * from '@vue-flow/minimap'
 export * from '@vue-flow/core'
